+++ server/Gateway.js
@@ -1,10 +1,11 @@
 class Gateway {
-  constructor(id, displayName, isActive, createdAt, updatedAt, client) {
+  constructor(id, displayName, isActive, createdAt, updatedAt, sentVitalsAlertAt, client) {
     this.id = id
     this.displayName = displayName
     this.isActive = isActive
     this.createdAt = createdAt
     this.updatedAt = updatedAt
+    this.sentVitalsAlertAt = sentVitalsAlertAt
     this.client = client
   }
 }